pyproject.toml
[tool.poetry]
name = "sca"
version = "0.1.0"
description = ""
authors = ["VAlduinV <prime72w@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
coreapi = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
